KillButton: Add "Kill as Admin" btn
If the run belongs to the logged-in user, then it'll
show "Kill" button. If not, then  "Kill As Admin" btn.
For now, pulpito-ng does not verify if user has admin
privileges. That check happens on t-api after the request is sent.

This commit also refreshes the mutation obj for kill-button,
upon button click. This fixes the issue of retrying to kill runs
by clicking the button again.

Signed-off-by: Vallari Agrawal <val.agl002@gmail.com>